location adjustment
stack.py location adjustment
